+++ R/check-vector-value.R
@@ -1,29 +1,30 @@
-check_vector_value_missing <- function(vector, value, column_name, data_name) {
+check_vector_value_missing <- function(vector, value, column_name, data_name = NULL) {
   if (!length(value) || !length(vector))
     return(TRUE)
   missing <- any(is.na(value))
   value <- value[!is.na(value)]
 
   if (!missing && any(is.na(vector)))
-    check_stop("column ", column_name, " in ", data_name, " cannot include missing values")
+      check_stop(name_info(column_name, data_name), " cannot include missing values")
 
   if (!length(value)) {
     if (!all(is.na(vector)))
-      check_stop("column ", column_name, " in ", data_name, " can only include missing values")
+        check_stop(name_info(column_name, data_name), " can only include missing values")
     return(TRUE)
   }
   value
 }
 
-check_vector_value <- function(vector, value, column_name, data_name)
+check_vector_value <- function(vector, value, column_name, data_name = NULL)
   UseMethod("check_vector_value")
 
-check_vector_value.default <- function(vector, value, column_name, data_name) {
+check_vector_value.default <- function(vector, value, column_name, data_name = NULL) {
   if (length(value) == 2) {
     range <- range(vector, na.rm = TRUE)
     value <- sort(value)
     if (range[1] < value[1] || range[2] > value[2])
-      check_stop("the values in column ", column_name, " in ", data_name, " must lie between ", value[1], " and ", value[2])
+        check_stop("the values in ", name_info(column_name, data_name),
+                   " must lie between ", value[1], " and ", value[2])
     return(TRUE)
   }
   if (!all(vector %in% value))
@@ -31,20 +32,20 @@ check_vector_value.default <- function(vector, value, column_name, data_name) {
   TRUE
 }
 
-check_vector_value.logical <- function(vector, value, column_name, data_name) {
+check_vector_value.logical <- function(vector, value, column_name, data_name = NULL) {
   value <- unique(value)
   if (length(value) == 2)
     return(TRUE)
   if (!all(vector == value))
-    check_stop("column ", column_name, " in ", data_name, " can only include ",
-              value, " values")
+      check_stop(name_info(column_name, data_name), " can only include ",
+                 value, " values")
   TRUE
 }
 
-check_vector_value.character <- function(vector, value, column_name, data_name) {
+check_vector_value.character <- function(vector, value, column_name, data_name = NULL) {
   if (length(value) == 2) {
     if (!all(grepl(value[1], vector, perl = TRUE) & grepl(value[2], vector, perl = TRUE)))
-      check_stop("column ", column_name, " in ", data_name, " contains strings that do not match both regular expressions ", punctuate(sort(value), qualifier = "and"))
+      check_stop(name_info(column_name, data_name), " contains strings that do not match both regular expressions ", punctuate(sort(value), qualifier = "and"))
     return(TRUE)
   }
   regexp <- paste0("(", paste(value, collapse = ")|(") , ")")
@@ -53,10 +54,10 @@ check_vector_value.character <- function(vector, value, column_name, data_name) 
   TRUE
 }
 
-check_vector_value.factor <- function(vector, value, column_name, data_name) {
+check_vector_value.factor <- function(vector, value, column_name, data_name = NULL) {
   if (length(value) == 2) {
     if (!all(as.character(value) %in% levels(vector)))
-      check_stop("column ", column_name, " in ", data_name, " lacks factor levels ", punctuate(sort(as.character(value)), qualifier = "and"))
+        check_stop(name_info(column_name, data_name), " lacks factor levels ", punctuate(sort(as.character(value)), qualifier = "and"))
     return(TRUE)
   }
   if (!identical(levels(value), levels(vector)))