+++ NEWS.md
@@ -7,6 +7,8 @@ to allow checking of the number of rows.
 - Added stricter variants of `check_data` called `check_data2` and `check_data3`
 as well as an alias `check_data1` for `check_data`.
 
+- Added function `check_vector` to check the class and values of a vector.
+- Added function `check_scalar` to check the class and values of a scalar.
 - Added function `check_data_frame` to check if an object is a data frame.
 - Added function `check_rows` to check the number of rows in a data frame.
 - Added function `check_cols` to check the names of columns in a data frame.